requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.3.5
requests==2.31.0
python-dateutil==2.8.2 